Seperated Code into Xcode and General Code
Seperated Code into Xcode and General Code
