--- GalagaClone/main.cpp
@@ -1,15 +1,0 @@
-//
-//  main.cpp
-//  GalagaClone
-//
-//  Created by Andrew Zimmerman on 4/3/18.
-//  Copyright © 2018 Andrew Zimmerman. All rights reserved.
-//
-
-#include <iostream>
-
-int main(int argc, const char * argv[]) {
-    // insert code here...
-    std::cout << "Hello, World!\n";
-    return 0;
-}
