+++ General Code/Bullet/Bullet.cpp
@@ -10,16 +10,17 @@
 
 #include "Bullet.h"
 
-
-flash::flash()
+// ## Begin Constructor ##
+Bullet::Bullet()
 {
     for(int i = 0; i<100; i++)
     {
         int j = 0;
         Shot[i].X = Shot[i].Y = 0;
-        Shot[i].X +=i;
-        Shot[i].Y +=j;
-        if(i%10 == 0){
+        Shot[i].X += i;
+        Shot[i].Y += j;
+        if(i % 10 == 0)
+        {
             j++;
         }
         color[i].R = 255;
@@ -28,37 +29,42 @@ flash::flash()
     }
 }
 
-void flash::setColor(int x,Color c)
+// ## Begin Color ##
+void Bullet::setColor(int x, Color c)
 {
-    color[x]=c;
+    color[x] = c;
 }
 
-void flash::setShot(Point p)
+Color Bullet::getColor(int x) const
+{
+    return color[x];
+}
+
+// ## Begin Shot ##
+void Bullet::setShot(Point p)
 {
     int m = 0, n = 0, x = 100;
-    for(int i = 0; i<x; i++){
-        Shot[i].X = p.X+m;
-        Shot[i].Y = p.Y+n;
+    for(int i = 0; i < x; i++)
+    {
+        Shot[i].X = p.X + m;
+        Shot[i].Y = p.Y + n;
         m++;
-        //***m%5, 5 is the wigth of bullet.
-        if((m%5) == 0){
+        // Wait time for Bullet is M % 5
+        if((m % 5) == 0)
+        {
             n++;
             m = 0;
         }
     }
 }
 
-Color flash::getColor(int x) const
-{
-    return color[x];
-}
-
-Point flash::getShot(int x) const
+Point Bullet::getShot(int x) const
 {
     return Shot[x];
 }
 
-void flash::displayFlash(int BulletSpeed, flash, SDL_Plotter& g)
+// ## Begin Display / Erase ##
+void Bullet::displayBullet(int BulletSpeed, Bullet myBullet, SDL_Plotter& g)
 {
     int x = 100;
     double m, n;
@@ -71,7 +77,7 @@ void flash::displayFlash(int BulletSpeed, flash, SDL_Plotter& g)
     }
 }
 
-void flash::eraseFlash(int BulletSpeed, flash, SDL_Plotter& g)
+void Bullet::eraseBullet(int BulletSpeed, Bullet myBullet, SDL_Plotter& g)
 {
     double m, n;
     int x = 100;
