+++ General Code/Ship/Ship.h
@@ -93,4 +93,5 @@ class Ship
         //************************************************************
         void display(SDL_Plotter&);
 };
+
 #endif /* Ship_H */
