+++ General Code/ShipFrame/ShipFrame.cpp
