--- README.md
@@ -1,73 +1,0 @@
-# 1430GalagaGroupProject
-A group project that attempts to make a spin off of the game Galaga using C++ and the SDL_Plotter framework.
-
-## Purpose
-To create a game using the SDL_Plotter framework that represents Galaga.
-
-## Details
-| Detail             | Info                       |
-|--------------------|----------------------------|
-| Status             | Under development          |
-| Date Created       | 3/29/2018                  |
-| Date Last Modified | 3/29/2018                  |
-| Language Used      | C++                        |
-| Frameworks used    | SDL_Plotter                |
-| Group Name         | Thought We Wouldn't Notice |
-
-## Contributors
-| Name             |
-|------------------|
-| Andrew Zimmerman |
-| Martyn Staalsen  |
-| Zixiao Li        |
-| Jiahao Xu        |
-| Evie Guay        |
-
-## Documentation Standards
-
-### Header Documentation
-
-        //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        //  Author: First Last
-        //  Assignment Title: Group Project
-        //  Assignment Description: Functions as a clone of Galaga.
-        //  Due Date: 4/26/2018
-        //  Date Created: MM/DD/YYYY
-        //  Date Last Modified: MM/DD/YYYY
-        //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        
-### If Contributions Were Made
-
-        //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        //  Author: First Last
-        //  Assignment Title: Group Project
-        //  Assignment Description: Functions as a clone of Galaga.
-        //  Due Date: 4/26/2018
-        //  Date Created: MM/DD/YYYY
-        //  Date Last Modified: MM/DD/YYYY
-        //  Contributor: First Last
-        //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-### Method Declarations
-
-        //************************************************************
-        // description:                                              *
-        // return:                                                   *
-        // precondition:                                             *
-        // postcondition:                                            *
-        //************************************************************
-        
-### Indicating Code Sections
-        // ## Segment Indication ##
-
-### Indicating What a Loop or If-Statement Does
-        while(some argument) // while ____ is true
-        { // perform action
-            ourAction();
-        }
-        
-        if(some argument) // if _____ is true
-        { // perform action
-            ourAction();
-        }
-
